Fix billing metric test failing because of no-datapoint-found
Sometimes, when we run a metrics query in OpenShift Monitoring, we
get "no datapoints found" as a result. If we wait a few minutes and
run the same query we get the result.

This commit adds retries to the "Run Query" keyword to avoid having
this issue

Also, as seen in this PR test screenshot, the keyword "Verify Query Results Contain Data" can fail
because the graph can show the text "no datapoints found" even when there is a correct returned
value. I think the graph needs more time to show the information
https://opendatascience-jenkins-csb-rhods.apps.ocp4.prod.psi.redhat.com/job/rhods-ci-pr-test/168/TestReport/selenium-screenshot-39.png

This commit also fixes the keyword not to check the existence of the referred string

Finally, this commit enhaces the billing metrics teardown

Signed-off-by: Jorge Garcia Oncins <jgarciao@redhat.com>