POC: Remove 'Library OpenShiftCLI'
The `OpenShiftCLI` library requires a valid `KUBECONFIG` to load
properly. Without it, the framework initialization fails because it
cannot access the cluster.

As part of our scale testing, we cannot allow any unnecessary call to
OpenShift, be it `oc login` or resource queries.

This commit is a quick-and-dirty POC showing that we can run our scale
test without relying on `OpenShiftCLI`.