[Fix] to distinguish between PVC in pending and PVC in bound status
The storage size is shown in the UI depending of the PVC status. Before
the PVC is bounded, there is shown 'Max XXGi' (or other unit). After the
PVC has been bound (even after it has been de-attached from a workbench)
then there is something like: 'XXGi [##     ] YYGi' - so there is shown
both the capacity that is taken and total capacity of the storage; in
the middle of them, there is a graphical element to show how much the
capacity is used.