[Fix] the test that checks the User/Group privileges permissions
This fixes the `Verify Unauthorized User Is Not Able To Spawn Jupyter Notebook`
test. It is basically similar change as done in [1,2].

Note: looks like the behavior of the RHOAI is a bit different now - the
user without permissions is shown directly the `Access permissions
needed` page. Nothing else is shown. Based on the test structure, in
previous versions we could see the application but weren't able to start
a JupyterHub tile in the `Applications -> Enabled` menu.

Also, unnecessary check for the product version is removed with this
commit and a small typo fix.

[1] https://github.com/red-hat-data-services/ods-ci/pull/1034
[2] c719b70d93c88734288ae064fefc71991ccb236a