Revert "[Fix] attempt to stabilize the `Logged Into OpenShift` test (#1065)"
This reverts commit 1cb3109542f44aa129b163dc26fdef79f4fcb84e.

Reason for revert is that it caused issue for Perf testing in the way
that the JupyterLab IDE hasn't been loaded properly (splash screen was
present on the screenshots from the test execution) and this check
returned a failure after cca 15 seconds.

We will have to find a different approach to improve on this in case
we'll see the problem in the IDE load in the future.