[Fix] wait for the RHOAI/ODH operator CSV to be in state condition Succeeded
The installation of ODH often failed due to the fact that we tried to
apply the DSC on the cluster even before such CRD resource is known to
the cluster (that means that operator itself hasn't been installed
completely yet). This change should wait for the installation to be
prepared for the DSC application. Timeout is set to 2 minutes.