Decrease TTL for our DNS records for the cluster creation.
At the moment we have 2days TTL for the DNS records for the cluster IPs.
Since this causes us troubles when we want to recreate the cluster with
the very same name (the records is probably still in the DNS cache of
the system where the new cluster is being created), let's reduce this
TTL value to 5minutes only.

The overhead with the DNS shouldn't be hopefully that bad.