Skip a PV Storage test on disconnected
Let's skip this test on Disconnected environment because we have a PVs predefined
there hardcoded to 100Gi. Since there is a size check in this test, it would fail
unless some extra condition to be brought in here. I believe that it is okay to
check just on regular cluster for now and skip on disconnected safely.

Note, PV and PVC is always 1-to-1 connection [1]. We can improve our
disconnected infrastructrue so that the dynamic PV provisioning is
enabled and then we can re-enable this test back.

* [1] https://kubernetes.io/docs/concepts/storage/persistent-volumes/#binding