[Fix] image creation via Dockerfile after move to Python 3.11
This is a followup of [1,2] and fixes image build creation after we
moved from Python 3.9 to Python 3.11

On top of that, it also removes dnf cache after all installation is
done. There is no change in the sice of resulting image but it's a
standard step to do in the Dockerfiles in general.

[1] #1118
[2] 867a617bc224726cf98fa3354293f8e50b4f5eb5