[Fix] Notebook API test
This test was probably failing since the addition of the accelerator
profiles. Since I have no idea where is some documentation to the
dashboard API which we use here, I grepped my local disk for some
keywords and was able to find this file [1] from which I was able to
understand what is wrong with this test. As such, please review
thoroughly since I could miss something.

[1]
https://github.com/opendatahub-io/odh-dashboard/blob/main/frontend/src/types.ts#L670>