[fix] of the get_osd_cluster_id function
We are trying to retrieve the ID of the cluster that OCM uses for its
work. For this we use either the cluster name or the cluster ID
(different ID to the one that OCM uses). The cluster ID is saved in the
external_id key.

Also, to keep this function reflexive, let's keep also `id` in the query
too. With this, we can always call this function and get the ocm ID for
the cluster anytime we use cluster `name`, `external ID` or `ID`.