[fix] Update image used for must-gather test
A new image was build accidentally a couple of days ago. This image
should contain same code as the previous one, but since it is build
recently, multiple packages from base image are updated.

Also, since this image has been marked for 2.9.0 already and is tagged
as stable, we should start to use it in our tests too [1,2].

[1] https://github.com/red-hat-data-services/rhoai-disconnected-install-helper/blob/main/rhoai-2.9.md
[2] https://access.redhat.com/solutions/7061604