Check CSV version before and after upgrade
Refactor upgrade test to check if version after upgrade is greater than the initial version. The previous approach that checked if the pods in the operator namespace were Pending after 10s was causing the test to fail, because the pods were usually already running before 10s. I think the right way to confirm if the upgrade was successful is to check the CSV version and that the pods in the operator namespace are running once the CSV version has been updated

rh-pre-commit.version: 2.2.0
rh-pre-commit.check-secrets: ENABLED