[fix] the test for special usernames
This started failing after this workaround #1432 ([1]). Originally the
test expected that after logout and another login, the user is
redirected to the original page where logout started. This doesn't
happen anymore because of this new redirect to `enabled` paged in the
mentioned workaround.

Once the workaround is removed, we can delete the line added in this PR
also. Though, the main point of the test isn't to check where the page
is open after login, but the actual login names work as expected only.

* [1] f9dfc88c6e9eba212b06f72eb210e17312967b8c