Verify RHODS version before and after upgrade (#1468)
* Verify RHODS version before and after upgrade

Refactor upgrade test to check if version after upgrade is greater than the initial version. The previous approach that checked if the pods in the operator namespace were Pending after 10s was causing the test to fail, because the pods were already running before 10s, and playing with that sleep timer might make the test flaky. I think the right way to confirm if the upgrade has been successful is to check the CSV version and that the pods in the operator namespace are running once the CSV version has been updated

rh-pre-commit.version: 2.2.0
rh-pre-commit.check-secrets: ENABLED

* Verify that operator pod creation date changes after upgrade

* Fix formatting