[fix] must-gather test to use the image targeted for 2.8.3
There was, yet again, accidentally build a new stable image of
must-gather meaning that the image was updated. Technically no changes
got in, but since its a new build, some packages and dependencies may be
of a newer version than before.

Anyway, it was decided that we're gonna use this for the release
unfortunatelly...