[fix] intermittent test issue for ODS-1883
The test `Verify User Can Create Environment Variables By Uploading YAML
Secret/ConfigMap` would fail quite often since we wanted to click on the
`Workbenches` tab before it is showed on the screen. Anyway, since the
original bug where dashboard was always moving back to the `Overview`
tab after a modal window was closed, it's not necessary to move to the
`Workbenches` tab expclicitly anymore, so removing this line.

For the record, this tab switch was added in [1,2].

* [1] https://github.com/red-hat-data-services/ods-ci/pull/1363
* [2] 556678d21fa928ff68733e695e081a4816785678