[fix] close browser after each test in 415__ods_dashboard_projects.robot
Since each tests in this test-case starts a new browser window, let's be
sure we also close it at the end of each test.

Motivation to this is an attempt to stabilize the Sanity run, where we
bumped into an issue when test execution stopped for some reason at the
execution of the: `Verify Event Log Is Accessible While Starting A Workbench`
test. I'm not sure whether this change helps but it definitely didn't
make things worse.