[fix] stabilize the ODS-1864 test by scrolling button into view
This stabilizes the `Verify User Can Create A Workbench With Environment
Variables` test - it usually pass just fine locally, but fails on
Jenkins since the button isn't clickable as it's not in the view of the
browser.

This tries to move the button into the center of the view.

Also, unnecessary version check removed.