Check that the URL list is not empty before verifying HTTP status
KW `URLs HTTP Status Code Should Be Equal To` receives a list of links
to verify that no link is broken. However if the list is empty
(most likely due to invalid XPATH), then no URL validation is executed.
This would cause a false positive (as if all links are working), instead
of failing the calling test (expected link elements are missing).

Signed-off-by: manosnoam <nmanos@redhat.com>