Fix KW `Clean Up DSP Page`
This is required with the recent change of RHOAIENG-9576:
Remove the DS Projects List Toggle

The KW `Clean Up DSP Page` should not loop on all projects in UI,
but only on those which have DS project annotation in its namespace.
In addition, fix the DS Project title element xpath.

Signed-off-by: manosnoam <nmanos@redhat.com>