Introduce support for the Jupyter/Elyra 4.
These changes are required to support the Jupyter/Elyra 4.x IDE version
bump which is expected in the OOTB Workbench images 2024.2+. All changes
are done so that the backward compatibility with the current
Jupyter/Elyra 3.x version is preserved.

Note: since the 'default' version of the workbench images still points
to 2024a with Jupyter 3, the line with the global env variable update to
work with Jupyter 4 is commented out now with a TODO comment.

Also, the reason why we are adding some code from the JupyterLibrary is
that we're trying to use SeleniumLibrary for the browser opening. This
doesn't match with some part of the code of the JupyterLibrary where it
is expected that some keywords that are present in both (SeleniumLibrary
and JupyterLibrary) are used from the JupyterLibrary. Since to narrow
down this discrepancy (using JupyterLibrary explicitly and completely
for the JupyterHub testing) would require more changes, I did it via
this ughly hack now having part of the source from the JupyterLibrary
directly in our codebase.
Truth is that JupyterLibrary project seems to be abandonded so we may
also think about fork it and handle it by ourselves in case we'll need
more changes there.