Fix for the Elyra pipeline execution
The name of the pipeline created depended on the name of the image. So
if we used a different or custom image (e.g. during testing some new
image) the value generated from `${img} Pipeline` didn't match the
expected name of the pipeline experiment later in the code. It was
wrong from the beginning and the name of the experiment should be used
also here.

Works with the current Jupyter/Elyra 3 image and also with the upcoming
Jupyter/Elyra 4 image.