[Fix] improve the way we wait for the particular Dashboard page to load
For some dashboard pages, it takes time to load properly and they can
change the content during that time. E.g. there is a "Loading" text with
an animated gif provided until the actual page is present.

In case of accessing the Applications -> Enabled -> Jupyter app tile,
there may be also present a page title with "Loading..." text for a
brief amount of time. Such thing breaks our detection technique as then
our expected page title text doesn't match with what was present on the
page and fails with:
```
BuiltIn.Should Be Equal:
Loading... != Start a notebook server
```

Adding a couple of retries should help to avoid such situation.