[Fix] the test for the package versions on images
In the second part of this test, we have a selected list of packages so
that we expect these packages to be same version amongst all the tested
images. We have to reduce this list as now there are more packages that
are expected not to be in the same version as was in past.