Update the way how to remove environment variables for Jupyter ntb
This change slightly simplifies the way how the variables are removed
from the Jupyter spawner page for the notebook - now it doesn't depend
on particular variable values, but simply clicks until the remove button
is present on the page.

Another change here is that this variable removal is now run also in
case that the given set of variables to the `Spawn Notebook With
Arguments` keyword is empty. This will assure that any eventual
leftovers from previous test are cleaned up properly not affecting this
current test.