Deploy SR-IOV operator and update ClusterPolicy for RDMA (#2668)
* Increase timeouts for GPU and NFD installation

I've noticed a few times that the subscription status isn't reached
in the given timeouts but after waiting and re-running the script it
works as expected, signaling that the timeout in some cases is too
low.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

* Deploy SR-IOV operator and update ClusterPolicy for RDMA

for testing llm-d, we might need to use SR-IOV and the Nvidia kernel
parameters `EnableStreamMemOPs=1` and `PeerMappingOverride=1`.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

* Revert "Increase timeouts for GPU and NFD installation"

This reverts commit 7e7f47b6d8ddaf2147f562346e4e3edb2666f2b0.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

---------

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>
Co-authored-by: Berto D'Attoma <bdattoma@redhat.com>